Create docker images which speed up MM generation.
The trick here is, every time Magic Modules gets a new commit merged in,
we'll generate some new docker images based on that commit.  The idea
is, we'll create images which already have the necessary dependencies
installed (most of the time).  We'll still run `go get` in the
generate-terraform script, but ideally it won't turn out to grab
anything, because the container will already have everything it needs.
Same story with the other generators - we'll run `bundle install`
beforehand so that we can save some time since everything will be there
already.